% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.model.R
\name{cv.model}
\alias{cv.model}
\title{idm cv model update of beta in penalised regression}
\usage{
cv.model(
  beta,
  nva01,
  nva02,
  nva12,
  fix,
  penalty.factor,
  penalty,
  v,
  fu,
  lambda,
  alpha
)
}
\arguments{
\item{beta}{parameters on explanatory variables}

\item{nva01}{number of variables for transition 0 -->1}

\item{nva02}{number of variables for transition 0 -->2}

\item{nva12}{number of variables for transition 1 -->2}

\item{fix}{indicators of fixed and unfixed parameters}

\item{penalty.factor}{base::which variable should be penalised}

\item{penalty}{base::which penalty to consider}

\item{v}{variance covariance matrix}

\item{fu}{-loglikelihood}

\item{lambda}{lambda penalised parameter}

\item{alpha}{alpha penalised parameter}
}
\description{
idm cv model update of beta in penalised regression
}
\author{
R: Ariane Bercu, Celia Touraine <Celia.Touraine@isped.u-bordeaux2.fr> and Thomas Alexander Gerds <tag@biostat.ku.dk>
Fortran: Pierre Joly <Pierre.Joly@isped.u-bordeaux2.fr>
}
