% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivaweib.R
\name{derivaweib}
\alias{derivaweib}
\title{First and second order derivatives of the log-likelihood with weibull baseline risk}
\usage{
derivaweib(
  b,
  npm,
  npar,
  bfix,
  fix,
  ctime,
  no,
  ve01,
  ve02,
  ve12,
  dimnva01,
  dimnva02,
  dimnva12,
  nva01,
  nva02,
  nva12,
  t0,
  t1,
  t2,
  t3,
  troncature
)
}
\arguments{
\item{b}{parameters on explanatory variables not fixed}

\item{npm}{number of parameters not fixed, thus length of b}

\item{npar}{number of parameters}

\item{bfix}{parameters on explanatory variables fixed}

\item{fix}{indicator of length npar, values 1 if parameter fixed}

\item{ctime}{profile of patients from 1 to 7}

\item{no}{number of subjects}

\item{ve01}{variables for transition 0 -->1}

\item{ve02}{variables for transition 0 -->2}

\item{ve12}{variables for transition 1 -->2}

\item{dimnva01}{number of variables for transition 0 -->1, if not variables value 1}

\item{dimnva02}{number of variables for transition 0 -->2, if not variables value 1}

\item{dimnva12}{number of variables for transition 1 -->2, if not variables value 1}

\item{nva01}{number of variables for transition 0 -->1, if not variables value 0}

\item{nva02}{number of variables for transition 0 -->2, if not variables value 0}

\item{nva12}{number of variables for transition 1 -->2, if not variables value 0}

\item{t0}{time of entry}

\item{t1}{time of last visit or last visit without diagnose of illness}

\item{t2}{time of last visit or time diagnose of illness}

\item{t3}{time of last visit or death}

\item{troncature}{indicator of troncature, value 1 if there is troncature otherwise 0.}
}
\description{
First and second order derivatives of the log-likelihood with weibull baseline risk
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
