% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idm.penalty.R
\name{idm.penalty}
\alias{idm.penalty}
\title{programmation for penalised idm with splines as risk bases}
\usage{
idm.penalty(
  b,
  fix0,
  size_V,
  size_spline,
  clustertype,
  epsa,
  epsb,
  epsd,
  eps.spline,
  eps.eigen,
  print.info,
  nproc,
  maxiter,
  maxiter.pena,
  nfolds,
  knots01,
  knots02,
  knots12,
  ctime,
  N,
  nknots01,
  nknots02,
  nknots12,
  ve01,
  ve02,
  ve12,
  dimnva01,
  dimnva02,
  dimnva12,
  nvat01,
  nvat02,
  nvat12,
  t0,
  t1,
  t2,
  t3,
  troncature,
  gauss.point,
  nlambda01,
  lambda01,
  nlambda02,
  lambda02,
  nlambda12,
  lambda12,
  alpha,
  penalty.factor,
  step.sequential,
  option.sequential,
  penalty
)
}
\description{
programmation for penalised idm with splines as risk bases
}
