% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idm.weib.R
\name{idm.weib}
\alias{idm.weib}
\title{Illness-death model algorithm with weibull baseline risk}
\usage{
idm.weib(
  b,
  fix0,
  size_V,
  clustertype,
  epsa,
  epsb,
  epsd,
  nproc,
  maxiter,
  ctime,
  N,
  ve01,
  ve02,
  ve12,
  dimnva01,
  dimnva02,
  dimnva12,
  nvat01,
  nvat02,
  nvat12,
  t0,
  t1,
  t2,
  t3,
  idd,
  idm,
  ts,
  troncature,
  gausspoint,
  weib,
  methodCV
)
}
\arguments{
\item{b}{parameters not fixed}

\item{fix0}{indicators of fixed and unfixed parameters}

\item{size_V}{number of parameters}

\item{clustertype}{in which cluster to work}

\item{epsa}{control convergence parameter for beta}

\item{epsb}{control convergence parameter for loglik}

\item{epsd}{control convergence for distance to minimum rdm}

\item{nproc}{number of cluster}

\item{maxiter}{Maximum number of iterations. The default is 200.}

\item{ctime}{classification of subject according to their observations}

\item{N}{number of subjects}

\item{ve01}{variables for transition 0 -->1}

\item{ve02}{variables for transition 0 -->2}

\item{ve12}{variables for transition 1 -->2}

\item{dimnva01}{number of variables for transition 0 -->1}

\item{dimnva02}{number of variables for transition 0 -->2}

\item{dimnva12}{number of variables for transition 1 -->2}

\item{nvat01}{number of variables for transition 0 -->1}

\item{nvat02}{number of variables for transition 0 -->2}

\item{nvat12}{number of variables for transition 1 -->2}

\item{t0}{time entry}

\item{t1}{time L}

\item{t2}{time R}

\item{t3}{time of event/out}

\item{idd}{number of subjects that died}

\item{idm}{number of subjects that had illness}

\item{ts}{delay in the study}

\item{troncature}{indicator if troncature or not}

\item{gausspoint}{number of points in gauss quadrature}

\item{weib}{the form of the weibull parameters}

\item{methodCV}{methodolgy to optimise penalised parameters}
}
\description{
Illness-death model algorithm with weibull baseline risk
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
