% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.idm.R
\name{predict.idm}
\alias{predict.idm}
\title{Predict transition probabilities and cumulative probabilities from an object
of class \code{idmSplines} with confidence intervals are calculated.}
\usage{
\method{predict}{idm}(
  object,
  s,
  t,
  newdata,
  nsim = 200,
  seed = 21,
  conf.int = 0.95,
  lifeExpect = FALSE,
  maxtime,
  lambda = "BIC",
  ...
)
}
\arguments{
\item{object}{an \code{idm} class objects returned by a call to the
\code{\link{idm}} function}

\item{s}{time point at which prediction is made.}

\item{t}{time horizon for prediction.}

\item{newdata}{A data frame with covariate values for prediction.}

\item{nsim}{number of simulations for the confidence intervals
calculations.  The default is 200.}

\item{seed}{Seed passed to \code{set.seed} for Monte Carlo
simulation of confidence intervals.}

\item{conf.int}{Level of confidence, i.e., a value between 0 and 1,
the default is \code{0.95}.  The default is also used when
\code{conf.int=TRUE}.  To avoid computation of confidence
intervals, set \code{conf.int} to FALSE or NULL.}

\item{lifeExpect}{Logical. If \code{TRUE} compute life
expectancies, i.e., \code{t=Inf}.}

\item{maxtime}{The upper limit of integration for calculations of life expectancies from Weibull parametrizations.}

\item{lambda}{If penalised illness-death model define if prediction need to be performed 
looking at the minimum of BIC or GCV or a vector of three values for each transition 0 -> 1, 0 -> 2 and 1 -> 2.}

\item{...}{other parameters.}
}
\value{
a list containing the following predictions with pointwise
    confidence intervals: \item{p00}{the transition probability
    \eqn{p_{00}}.}  \item{p01}{the transition probability
    \eqn{p_{01}}.} \item{p11}{the transition probability
    \eqn{p_{11}}.} \item{p12}{the transition probability
    \eqn{p_{12}}.} \item{p02_0}{the probability of direct
    transition from state 0 to state 2.} \item{p02_1}{the
    probability of transition from state 0 to state 2 via state 1.}
    \item{p02}{transition probability \eqn{p_{02}}. Note that
    \code{p02}=\code{p_02_0}+\code{p02_1}.} \item{F01}{the lifetime
    risk of disease. \code{F01}=\code{p01}+\code{p02_1}.}
    \item{F0.}{the probability of exit from state
    0. \code{F0.}=\code{p02_0}+\code{p01}+\code{p02_1}.}
}
\description{
Predictions for an illness-death model using either a penalized likelihood
approach or a Weibull parametrization.
}
\examples{

\dontrun{
library(lava)
library(prodlim)
set.seed(17)
d <- simulateIDM(n=1000,
                 beta01=c(1,1,0,0.5,0.5,rep(0,5)),
                 beta02=c(1,0,0,0,0.5,rep(0,5)),
                 beta12=c(1,0,0,0,0.5,rep(0,5)))$data
                 
fitpenweib <- idm(formula01=Hist(time=list(L,R),event=seen.ill)~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                  formula02=Hist(time=observed.lifetime,event=seen.exit)~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                  formula12=~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                  data=d,penalty="lasso",lambda01 = c(10,20),lambda02 = 10, lambda12 = 10)
predict(fitpenweib,s=10,t=15,lambda = "BIC") 
}
}
\seealso{
\code{\link{idm}}
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
\keyword{methods}
