% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateIDM.R
\name{simulatepenIDM}
\alias{simulatepenIDM}
\title{Sample illness-death model data}
\usage{
simulatepenIDM(
  n = 100,
  seed,
  scale.illtime,
  shape.illtime,
  scale.lifetime,
  shape.lifetime,
  scale.waittime,
  shape.waittime,
  prob.censoring,
  administrative.censoring,
  n.inspections,
  schedule,
  punctuality,
  nvar,
  mean,
  sd,
  cov,
  x01,
  x02,
  x12,
  beta01,
  beta02,
  beta12
)
}
\arguments{
\item{n}{number of observations}
}
\description{
Simulate data from an illness-death model with interval censored event times and penalized covariates
}
\details{
Simulate data from an illness-death model with interval censored event times
and covariates for the purpose of illustrating the help pages of the SmoothHazard package.
See the body of the function for details, i.e., evaluate simulateIDM
}
\seealso{
idmModel sim.idmModel simulateIDM
}
