% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothHazardoptim9-package.R
\docType{package}
\name{SmoothHazardoptim9-package}
\alias{SmoothHazardoptim9}
\alias{SmoothHazardoptim9-package}
\title{SmoothHazardoptim9: Estimation of Smooth Hazard Models for Interval-Censored Data with Applications to Survival and Illness-Death Models}
\description{
Estimation irreversible illness-death models with possibly interval-censored, left-truncated and right-censored data. Proportional intensities regression models can be specified to allow for covariates effects separately for each transition. With a limited number of covariates, we use either a parametric approach with Weibull or M-spline for baseline intensities. Parameter estimates are obtained by maximum likelihood using Marquardt-Levenberg algorithm. With a large number of covariates, we apply a regularization on the likelihood, such as lasso, elastic-net, Minimax Concave Penalty or Smoothly clipped absolute deviation. Parameter estimation and selection are obtained by maximising the penalised log-likelihood using a proximal gradient algorithm.
}
\keyword{internal}
