% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinesMI.R
\name{splinesMI}
\alias{splinesMI}
\title{@title M-spline extracted from fortran coding in package}
\usage{
splinesMI(x, knots, Boundary.knots)
}
\arguments{
\item{x}{vector in which splines will be evaluated}

\item{knots}{internal nodes values}

\item{Boundary.knots}{external nodes values}
}
\description{
@title M-spline extracted from fortran coding in package
}
\examples{
\dontrun{
splinesMI(x=seq(0.01,10,length.out=20),knots=c(2,5,8),Boundary.knots=c(0.01,10))$Mspline
}
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
\keyword{methods}
