% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.idm.R
\name{summary.idm}
\alias{summary.idm}
\title{Summary of a fitted illness-death model}
\usage{
\method{summary}{idm}(object, conf.int = 0.95, digits = 4, pvalDigits = 4, eps = 1e-04, ...)
}
\arguments{
\item{object}{a \code{idmSplines} object, i.e., the result of a call to the
\code{\link{idm}} function with \code{intensities}="Splines".}

\item{conf.int}{The level of confidence for the hazard ratios. The default is \code{0.95}.}

\item{digits}{number of digits to print.}

\item{pvalDigits}{number of digits to print for p-values.}

\item{eps}{convergence criterion used for p-values.}

\item{\dots}{other unusued arguments.}
}
\description{
Summarize the event history data of an illness-death regression model
and show regression coefficients for transition intensities
}
\examples{

\dontrun{
library(lava)
library(prodlim)
set.seed(17)
d <- simulateIDM(n=1000,
                 beta01=c(1,1,0,0.5,0.5,rep(0,5)),
                 beta02=c(1,0,0,0,0.5,rep(0,5)),
                 beta12=c(1,0,0,0,0.5,rep(0,5)))$data
                 
fitpenweib <- idm(formula01=Hist(time=list(L,R),event=seen.ill)~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                  formula02=Hist(time=observed.lifetime,event=seen.exit)~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                  formula12=~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                  data=d,penalty="lasso",lambda01 = c(10,20),lambda02 = 10, lambda12 = 10)
summary(fitpenweib)
}
}
\seealso{
\code{\link{idm}}, \code{\link{print.idm}},
\code{\link{plot.idm}}
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
\keyword{methods}
