% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmlLikelihood.R
\name{idmlLikelihood}
\alias{idmlLikelihood}
\title{idm log likelihood}
\usage{
idmlLikelihood(
  b,
  np,
  npar,
  bfix,
  fix,
  zi01,
  zi02,
  zi12,
  ctime,
  no,
  nz01,
  nz02,
  nz12,
  ve01,
  ve02,
  ve12,
  dimnva01,
  dimnva02,
  dimnva12,
  nva01,
  nva02,
  nva12,
  t0,
  t1,
  t2,
  t3,
  troncature,
  gausspoint
)
}
\arguments{
\item{b}{parameters not fixed}

\item{np}{number of parameters not fixed}

\item{npar}{number of parameters}

\item{bfix}{parameters fixed}

\item{fix}{indicators of fixed and unfixed parameters}

\item{zi01}{knots of transition 0 --> 1}

\item{zi02}{knots of transition 0 --> 2}

\item{zi12}{knots of transition 1 --> 2}

\item{ctime}{classification of subject according to their observations}

\item{no}{number of subjects}

\item{nz01}{number of knots for transition 0 -->1}

\item{nz02}{number of knots for transition 0 -->2}

\item{nz12}{number of knots for transition 1 -->2}

\item{ve01}{variables for transition 0 -->1}

\item{ve02}{variables for transition 0 -->2}

\item{ve12}{variables for transition 1 -->2}

\item{dimnva01}{number of variables for transition 0 -->1}

\item{dimnva02}{number of variables for transition 0 -->2}

\item{dimnva12}{number of variables for transition 1 -->2}

\item{nva01}{number of variables for transition 0 -->1}

\item{nva02}{number of variables for transition 0 -->2}

\item{nva12}{number of variables for transition 1 -->2}

\item{t0}{time entry}

\item{t1}{time L}

\item{t2}{time R}

\item{t3}{time of event/out}

\item{troncature}{indicator if troncature or not}

\item{gausspoint}{number of gausspoint quadrature}
}
\description{
idm log likelihood
}
\author{
R: Ariane Bercu, Celia Touraine <Celia.Touraine@isped.u-bordeaux2.fr> and Thomas Alexander Gerds <tag@biostat.ku.dk>
Fortran: Pierre Joly <Pierre.Joly@isped.u-bordeaux2.fr>
}
