% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmlLikelihood.R
\name{idmlLikelihood}
\alias{idmlLikelihood}
\title{idm log likelihood}
\usage{
idmlLikelihood(
  b0,
  np0,
  npar0,
  bfix0,
  fix0,
  zi010,
  zi020,
  zi120,
  c0,
  no0,
  nz010,
  nz020,
  nz120,
  ve010,
  ve020,
  ve120,
  dimnva01,
  dimnva02,
  dimnva12,
  nva01,
  nva02,
  nva12,
  t00,
  t10,
  t20,
  t30,
  troncature0,
  gausspoint0
)
}
\arguments{
\item{b0}{parameters not fixed}

\item{np0}{number of parameters not fixed}

\item{npar0}{number of parameters}

\item{bfix0}{parameters fixed}

\item{fix0}{indicators of fixed and unfixed parameters}

\item{zi010}{knots of transition 0 --> 1}

\item{zi020}{knots of transition 0 --> 2}

\item{zi120}{knots of transition 1 --> 2}

\item{c0}{classification of subject according to their observations}

\item{no0}{number of subjects}

\item{nz010}{number of knots for transition 0 -->1}

\item{nz020}{number of knots for transition 0 -->2}

\item{nz120}{number of knots for transition 1 -->2}

\item{ve010}{variables for transition 0 -->1}

\item{ve020}{variables for transition 0 -->2}

\item{ve120}{variables for transition 1 -->2}

\item{dimnva01}{number of variables for transition 0 -->1}

\item{dimnva02}{number of variables for transition 0 -->2}

\item{dimnva12}{number of variables for transition 1 -->2}

\item{nva01}{number of variables for transition 0 -->1}

\item{nva02}{number of variables for transition 0 -->2}

\item{nva12}{number of variables for transition 1 -->2}

\item{t00}{time entry}

\item{t10}{time L}

\item{t20}{time R}

\item{t30}{time of event/out}

\item{troncature0}{indicator if troncature or not}

\item{gausspoint0}{number of gausspoint quadrature}
}
\description{
idm log likelihood
}
\author{
R: Ariane Bercu, Celia Touraine <Celia.Touraine@isped.u-bordeaux2.fr> and Thomas Alexander Gerds <tag@biostat.ku.dk>
Fortran: Pierre Joly <Pierre.Joly@isped.u-bordeaux2.fr>
}
