% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity}
\alias{intensity}
\title{M-spline or Weibull estimate of the transition intensity function}
\usage{
intensity(
  times,
  knots,
  number.knots = NULL,
  theta,
  linear.predictor = 0,
  V = NULL,
  fix = NULL,
  converged = NULL,
  conf.int = 0.95,
  method = "weib"
)
}
\arguments{
\item{times}{Time points at which to estimate the intensity function}

\item{knots}{Knots for the M-spline}

\item{number.knots}{Number of knots for the M-splines (and I-splines see details)}

\item{theta}{The coefficients for the linear combination of M-splines (and I-splines see details), theta from the model}

\item{linear.predictor}{Linear predictor beta*Z. When it is non-zero,
transition and cumulative transition are multiplied by \code{exp(linear.predictor)}. Default is zero.}

\item{V}{matrix of variance co-variance}

\item{fix}{indicators if parameters are fixed}

\item{converged}{indicator of convergence of the models}

\item{conf.int}{confidence intervals, 1 - alpha}

\item{method}{the methodology, splines or weib}
}
\value{
\item{times}{The time points at which the following estimates are evaluated.}
\item{intensity}{The transition intensity function evaluated at \code{times}.}
\item{cumulative.intensity}{The cumulative transition intensity function evaluated at \code{times}}
\item{survival}{The "survival" function, i.e., exp(-cumulative.intensity)}
}
\description{
M-spline or Weibull estimate of the transition intensity function
and the cumulative transition intensity function
for survival and illness-death models
}
\details{
Important: the theta parameters returned by \code{idm} are in fact
the square root of the splines coefficients. See examples.
}
\examples{


\dontrun{
           
  library(lava)
library(prodlim)
set.seed(17)
d <- simulateIDM(n=1000)$data
fitweib <- idm(formula01=Hist(time=list(L,R),event=seen.ill)~X1+X2,
               formula02=Hist(time=observed.lifetime,event=seen.exit)~X1+X2,
               formula12=Hist(time=observed.lifetime,event=seen.exit)~X1+X2,data=d)
intensity(times=fitweib$time[,1],
               theta = fitweib$modelPar[1:2]^2)
fitsplines <- idm(formula01=Hist(time=list(L,R),event=seen.ill)~X1+X2,
                  formula02=Hist(time=observed.lifetime,event=seen.exit)~X1+X2,
                  formula12=Hist(time=observed.lifetime,event=seen.exit)~X1+X2,data=d,
                  method="splines")
intensity(times=fitsplines$time[,1],
               theta = fitsplines$theta01^2,
               knots = fitsplines$knots01,
               number.knots = fitsplines$nknots01,
               method = "splines")
}
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
