% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.predict.idm.R
\name{print.predict.idm}
\alias{print.predict.idm}
\title{Print prediction of illness-death model}
\usage{
\method{print}{predict.idm}(x, digits = 3, ...)
}
\arguments{
\item{object}{an \code{idm} class objects returned by a call to the
\code{\link{idm}} function}
}
\description{
Print prediction of illness-death model
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
