% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateIDM.R
\name{sim.idmModel}
\alias{sim.idmModel}
\title{Simulate illness-death model data}
\usage{
\method{sim}{idmModel}(
  x,
  n,
  plot,
  pen,
  illness.known.at.death = FALSE,
  latent = FALSE,
  keep.inspectiontimes = FALSE,
  semi.markov = T,
  ...
)
}
\arguments{
\item{x}{An \code{idmModel} object as obtained with
\code{idmModel}}

\item{n}{Number of observations}

\item{plot}{plot of base survival for all transition}

\item{pen}{if the simulated data are for a penalised version of illness-death model}

\item{illness.known.at.death}{Affects the value of variable
seen.ill}

\item{latent}{if TRUE keep the latent event times}

\item{keep.inspectiontimes}{if \code{TRUE} keep the inspection
times.}

\item{semi.markov}{True if you want to simulate 1 --> 2 in Semi Markov process}

\item{...}{Extra arguments given to \code{sim}}

\item{compliance}{Probability of missing an inspection time.}
}
\value{
A data set with interval censored observations from an illness-death model
example(idmModel)
help(idmModel)
}
\description{
Function to simulate illness-death model data
}
\details{
Based on the functionality of the lava PACKAGE
}
\author{
Thomas Alexander Gerds
}
