% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateIDM.R
\name{sim.idmModel}
\alias{sim.idmModel}
\title{Simulate illness-death model data}
\usage{
\method{sim}{idmModel}(x, n, plot, latent = FALSE, keep.inspectiontimes = FALSE, ...)
}
\arguments{
\item{x}{An \code{idmModel} object as obtained with
\code{idmModel}}

\item{n}{Number of observations
seen.ill}

\item{plot}{plot of base survival for all transition}

\item{latent}{if TRUE keep the latent event times}

\item{keep.inspectiontimes}{if \code{TRUE} keep the inspection
times.}

\item{...}{Extra arguments given to \code{sim}}

\item{compliance}{Probability of missing an inspection time.}
}
\value{
A data set with interval censored observations from an illness-death model
example(idmModel)
help(idmModel)
}
\description{
Function to simulate illness-death model data
}
\details{
Based on the functionality of the lava PACKAGE
}
\examples{
 \dontrun{
library(lava)
library(prodlim)
set.seed(17)
d <- simulateIDM(n=1000,
                 beta01=c(1,1,0,0.5,0.5,rep(0,5)),
                 beta02=c(1,0,0,0,0.5,rep(0,5)),
                 beta12=c(1,0,0,0,0.5,rep(0,5)))$data
}
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
