% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateIDM.R
\name{simulatepenIDM}
\alias{simulatepenIDM}
\title{Sample illness-death model data for penalised regression}
\usage{
simulatepenIDM(
  n = 100,
  seed,
  scale.illtime,
  shape.illtime,
  scale.lifetime,
  shape.lifetime,
  scale.waittime,
  shape.waittime,
  prob.censoring,
  administrative.censoring,
  n.inspections,
  schedule,
  punctuality,
  nvar,
  mean,
  sd,
  cov,
  x01,
  x02,
  x12,
  beta01,
  beta02,
  beta12,
  semi.markov = T
)
}
\arguments{
\item{n}{number of observations}

\item{seed}{specify the seed}

\item{scale.illtime}{Weilbull scale for latent illness time}

\item{shape.illtime}{Weilbull shape for latent illness time}

\item{scale.lifetime}{Weilbull scale for latent life time}

\item{shape.lifetime}{Weilbull shape for latent life time}

\item{scale.waittime}{Weilbull scale for latent life time}

\item{shape.waittime}{Weilbull shape for latent life time}

\item{prob.censoring}{probability of censoring at each visit}

\item{administrative.censoring}{specify time of administrative censoring}

\item{n.inspections}{Number of inspection times}

\item{schedule}{Mean of the waiting time between adjacent
inspections.}

\item{punctuality}{Standard deviation of waiting time between
inspections.}

\item{nvar}{number of variables}

\item{mean}{mean of each explanatory variables}

\item{sd}{standard-error of each explanatory variables}

\item{cov}{covariance matrix of explanatory variables}

\item{x01}{names of variables on transition 0 --> 1}

\item{x02}{names of variables on transition 0 --> 2}

\item{x12}{names of variables on transition 1 --> 2}

\item{beta01}{value of beta on transition 0 --> 1}

\item{beta02}{value of beta on transition 0 --> 2}

\item{beta12}{value of beta on transition 1 --> 2}

\item{semi.markov}{True if you want to simulate 1 --> 2 in Semi Markov process}
}
\description{
Simulate data from an illness-death model with interval censored event times and penalized covariates
}
\details{
Simulate data from an illness-death model with interval censored event times
and covariates for the purpose of illustrating the help pages of the SmoothHazard package.
See the body of the function for details, i.e., evaluate simulateIDM
}
\seealso{
idmModel sim.idmModel simulateIDM
}
