% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothHazardoptim9-package.R
\docType{package}
\name{SmoothHazardoptim9-package}
\alias{SmoothHazardoptim9}
\alias{SmoothHazardoptim9-package}
\title{SmoothHazardoptim9: Estimation of Smooth Hazard Models for Interval-Censored Data with Applications to Survival and Illness-Death Models}
\description{
Estimation of two-state (survival) models and irreversible illness- death models with possibly interval-censored,left-truncated and right-censored data. Proportional intensities regression models can be specified to allow for covariates effects separately for each transition. We use either a parametric approach with Weibull baseline intensities or a semi-parametric approach with M-splines approximation of baseline intensities in order to obtain smooth estimates of the hazard functions. Parameter estimates are obtained by maximum likelihood in the parametric approach and by penalized maximum likelihood in the semi-parametric approach.
}
\keyword{internal}
